+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add `original_value_container` property to `<al-field>` elements, for when they're not inside a `<al-form>` element but still need the original container
 * Use `getParentFieldElement()` instead of `alchemy_form` when resolving inherited attributes
+* Also let the `mode` attribute inherit
 
 ## 0.3.0-alpha.2 (2024-02-25)